increase the time span used for debug timestamps
git-svn-id: http://subversion.ffado.org/ffado/trunk/libffado@1188 2be59082-3212-0410-8809-b0798e1608f0

+++ src/debugmodule/debugmodule.cpp
@@ -173,12 +173,12 @@ DebugModule::print( debug_level_t level,
     // add a timing timestamp
     struct timespec ts;
     clock_gettime(CLOCK_MONOTONIC, &ts);
-    long unsigned int ts_usec=(uint32_t)(ts.tv_sec * 1000000LL + ts.tv_nsec / 1000LL);
+    uint64_t ts_usec=(uint64_t)(ts.tv_sec * 1000000LL + ts.tv_nsec / 1000LL);
 
     // format the message such that it remains together
     int chars_written=0;
     int retval=0;
-    retval = snprintf(msg, MB_BUFFERSIZE, "%010lu: %s (%s)[%4d] %s: ", 
+    retval = snprintf(msg, MB_BUFFERSIZE, "%011llu: %s (%s)[%4d] %s: ", 
                       ts_usec, getPreSequence( level ),
                       fname,  line,  function );
     if (retval >= 0) chars_written += retval; // ignore errors
